Chore: Organize date components
